fix: remove outdated cache test mode TODO comment (#472)
### **The Issue** 

There was a TODO comment in the cache code that said:

`// todo temp disable caching in test mode`

This comment was confusing because it made it seem like the caching
behavior in tests was incomplete or broken.

### **What this PR changes**

1. Removes the outdated TODO comment
2. No functional changes to the code
3. Just cleans up misleading documentation

### **Why I removed the TODO**

After investigating the codebase, I discovered that the current caching
behavior in test mode is intentionally designed and working correctly.